requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
zope.interface==5.5.0
